backend/requirements.txt
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.40.3
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
